# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AIMode": {
        "description": "Which engine actually produced the narrative in a response.\n\nExposed for the same reason as :class:`~app.schemas.common.DataMode`: a consumer\nmust always be able to tell real model output from a canned or fallback response.",
        "enum": [
          "gemini",
          "mock",
          "fallback"
        ],
        "title": "AIMode",
        "type": "string"
      },
...
      "Body_upload_crop_image_api_v1_farms__farm_id__crop_images_post": {
        "properties": {
          "farm_crop_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which planting this image belongs to.",
            "title": "Farm Crop Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "The crop photograph.",
            "title": "File",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional context for the diagnosis.",
            "title": "Note"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_crop_image_api_v1_farms__farm_id__crop_images_post",
        "type": "object"
      },
...
      "CropImage": {
        "description": "An uploaded crop image and its diagnosis, if one has run.",
        "properties": {
          "ai_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CropImageAnalysis"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null until analysis completes."
          },
          "analysis_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Error"
          },
          "analysis_status": {
            "$ref": "#/components/schemas/ImageAnalysisStatus"
          },
          "analyzed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          },
          "content_type": {
            "examples": [
              "image/jpeg"
            ],
            "title": "Content Type",
            "type": "string"
          },
          "farm_crop_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farm Crop Id"
          },
          "farm_id": {
            "format": "uuid",
            "title": "Farm Id",
            "type": "string"
          },
          "height": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional context supplied at upload.",
            "title": "Note"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signed URL for display. May expire.",
            "title": "Url"
          },
          "width": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "farm_id",
          "content_type",
          "size_bytes",
          "analysis_status",
          "uploaded_at"
        ],
        "title": "CropImage",
        "type": "object"
      },
...
      "CropImageList": {
        "description": "Response for `GET /api/v1/farms/{farm_id}/crop-images`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CropImage"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "CropImageList",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The one and only error shape returned by every 4xx/5xx response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "Farm": {
        "description": "Full server representation of a registered farm.",
        "properties": {
          "area_hectares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              12.5
            ],
            "title": "Area Hectares"
          },
          "country_code": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 3166-1 alpha-2, uppercase.",
            "examples": [
              "KE"
            ],
            "title": "Country Code"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crop_count": {
            "default": 0,
            "description": "Number of crops currently attached to this farm.",
            "minimum": 0.0,
            "title": "Crop Count",
            "type": "integer"
          },
          "elevation_m": {
            "anyOf": [
              {
                "maximum": 9000.0,
                "minimum": -500.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "farming_practice": {
            "$ref": "#/components/schemas/FarmingPractice",
            "default": "conventional"
          },
          "has_analysis": {
            "default": false,
            "description": "Whether an analysis run exists. Drives the dashboard empty state.",
            "title": "Has Analysis",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "irrigation_type": {
            "$ref": "#/components/schemas/IrrigationType",
            "default": "rainfed"
          },
          "latitude": {
            "examples": [
              -1.2921
            ],
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "examples": [
              36.8219
            ],
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "examples": [
              "North Field"
            ],
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Nairobi"
            ],
            "title": "Region"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "Farm",
        "type": "object"
      },
      "FarmCreate": {
        "description": "Request body for `POST /api/v1/farms`.",
        "properties": {
          "area_hectares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              12.5
            ],
            "title": "Area Hectares"
          },
          "country_code": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 3166-1 alpha-2, uppercase.",
            "examples": [
              "KE"
            ],
            "title": "Country Code"
          },
          "elevation_m": {
            "anyOf": [
              {
                "maximum": 9000.0,
                "minimum": -500.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "farming_practice": {
            "$ref": "#/components/schemas/FarmingPractice",
            "default": "conventional"
          },
          "irrigation_type": {
            "$ref": "#/components/schemas/IrrigationType",
            "default": "rainfed"
          },
          "latitude": {
            "examples": [
              -1.2921
            ],
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "examples": [
              36.8219
            ],
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "examples": [
              "North Field"
            ],
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Nairobi"
            ],
            "title": "Region"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude"
        ],
        "title": "FarmCreate",
        "type": "object"
      },
...
      "FarmList": {
        "description": "Response for `GET /api/v1/farms`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Farm"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "FarmList",
        "type": "object"
      },
...
      "ImageAnalysisStatus": {
        "enum": [
          "pending",
          "analyzing",
          "complete",
          "failed"
        ],
        "title": "ImageAnalysisStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "\n**TerraNex** is the central intelligence layer for agricultural decision support.\n\nIt fuses farm, crop, soil, weather and vegetation data with AI reasoning to produce\nfarm health analysis, weather and water risk, disease risk, agricultural advisories,\ncrop and regenerative-agriculture recommendations, and multimodal crop-image diagnosis.\n\n### For frontend developers\n\n* Every endpoint lives under `/api/v1`. Field casing is `snake_case` throughout.\n* Every failure returns the same envelope:\n  `{\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}`.\n  **Branch on `code`** \u2014 `message` is for humans and may change.\n* Every response carries `X-Request-Id`; include it when reporting a bug.\n* Generate typed bindings instead of hand-writing them:\n  `npx openapi-typescript ../contracts/openapi.json -o src/api/types.gen.ts`\n",
    "title": "TerraNex API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/crop-images/{image_id}": {
      "get": {
        "description": "Poll this while `analysis_status` is `pending` or `analyzing`.",
        "operationId": "get_crop_image_api_v1_crop_images__image_id__get",
        "parameters": [
          {
            "description": "Crop image identifier.",
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "description": "Crop image identifier.",
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropImage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "IMAGE_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Get a crop image",
        "tags": [
          "crop-images"
        ]
      }
    },
    "/api/v1/crop-images/{image_id}/analyze": {
      "post": {
        "description": "Runs multimodal analysis on a previously uploaded image, using the farm's crop and weather context to sharpen the diagnosis.\n\nThe result always states `is_plant_material`, lists `differential_diagnoses` rather than a single overconfident answer, and carries a `disclaimer`. Check `ai_mode` to see whether a real model or a mock produced it.",
        "operationId": "analyze_crop_image_api_v1_crop_images__image_id__analyze_post",
        "parameters": [
          {
            "description": "Crop image identifier.",
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "description": "Crop image identifier.",
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropImage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "IMAGE_NOT_FOUND"
          },
          "409": {
            "description": "ANALYSIS_IN_PROGRESS"
          },
          "503": {
            "description": "AI_UNAVAILABLE"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Diagnose a crop image",
        "tags": [
          "crop-images"
        ]
      }
    },
    "/api/v1/farms": {
      "get": {
        "operationId": "list_farms_api_v1_farms_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "List farms",
        "tags": [
          "farms"
        ]
      },
      "post": {
        "description": "Creates a farm at the given coordinates. Coordinates drive every downstream data lookup.",
        "operationId": "create_farm_api_v1_farms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Farm"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Register a farm",
        "tags": [
          "farms"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/crop-images": {
      "get": {
        "description": "Uploaded images with their diagnoses, newest first.",
        "operationId": "list_crop_images_api_v1_farms__farm_id__crop_images_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropImageList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "List a farm's crop images",
        "tags": [
          "crop-images"
        ]
      },
      "post": {
        "description": "Accepts `multipart/form-data`. Max 10 MB; allowed types: image/jpeg, image/png, image/webp.\n\nReturns immediately with `analysis_status: \"pending\"` so the UI can render the thumbnail at once. Pass `analyze=true` to run the diagnosis in the same request.",
        "operationId": "upload_crop_image_api_v1_farms__farm_id__crop_images_post",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "description": "Run the diagnosis immediately instead of returning pending.",
            "in": "query",
            "name": "analyze",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run the diagnosis immediately instead of returning pending.",
              "title": "Analyze",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_crop_image_api_v1_farms__farm_id__crop_images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropImage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "413": {
            "description": "IMAGE_TOO_LARGE"
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Upload a crop image",
        "tags": [
          "crop-images"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/recommendations/crops": {
      "get": {
        "description": "Catalog crops ranked by deterministic suitability against this farm's soil and climate, with the AI's reasoning attached to each.",
        "operationId": "get_crop_recommendations_api_v1_farms__farm_id__recommendations_crops_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "description": "How many to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "description": "How many to return.",
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropRecommendationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or NO_ANALYSIS_YET"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Crop recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/recommendations/regenerative": {
      "get": {
        "description": "Regenerative practices scored against this farm's soil organic carbon, texture and current farming practice, with expected soil-carbon and water-retention effects.",
        "operationId": "get_regenerative_recommendations_api_v1_farms__farm_id__recommendations_regenerative_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegenerativeRecommendationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or NO_ANALYSIS_YET"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Regenerative agriculture recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  }
}